requirements.txt
langchain
langchain-groq
pandas
streamlit
python-dotenv
PyPDF2
pydantic
